+++ app/src/main/java/com/example/wakeup/ui/main/database/viewmodels/utils/ApiHandler.java
@@ -1,4 +1,4 @@
-package com.example.wakeup.ui.main.utils;
+package com.example.wakeup.ui.main.database.viewmodels.utils;
 
 import java.util.HashMap;
 import java.util.Map;
