+++ app/src/main/java/com/example/wakeup/ui/main/database/viewmodels/utils/command/AddCommand.java
@@ -1,16 +1,15 @@
-package com.example.wakeup.ui.main.utils.command;
+package com.example.wakeup.ui.main.database.viewmodels.utils.command;
 
 import com.example.wakeup.ui.main.database.viewmodels.TaskViewModel;
 import com.example.wakeup.ui.main.models.Task;
 
-public class EditCommand extends Command{
-    public EditCommand(Task task){
+public class AddCommand extends Command{
+    public AddCommand(Task task){
         super(task);
     }
 
     @Override
     public void undo(TaskViewModel taskViewModel) {
-        taskViewModel.update(task);
+        taskViewModel.delete(task);
     }
-
 }
