+++ app/src/main/java/com/example/wakeup/ui/main/database/viewmodels/utils/command/EditCommand.java
@@ -1,15 +1,16 @@
-package com.example.wakeup.ui.main.utils.command;
+package com.example.wakeup.ui.main.database.viewmodels.utils.command;
 
 import com.example.wakeup.ui.main.database.viewmodels.TaskViewModel;
 import com.example.wakeup.ui.main.models.Task;
 
-public class AddCommand extends Command{
-    public AddCommand(Task task){
+public class EditCommand extends Command{
+    public EditCommand(Task task){
         super(task);
     }
 
     @Override
     public void undo(TaskViewModel taskViewModel) {
-        taskViewModel.delete(task);
+        taskViewModel.update(task);
     }
+
 }
